Refactor HomePage and ProductPage to utilize new component structure
- Updated HomePage and ProductPage classes to integrate HeaderComponent, FooterComponent, and StorefrontContentComponent.
- Removed unused methods and abstracted common functionalities into components.
- Introduced ProductComponent to encapsulate product-related functionalities.
- Enhanced test classes to align with the new component structure, ensuring proper interactions with the header and product components.
- Added new FooterComponent and StorefrontContentComponent classes as placeholders for future development.
