+++ src/main/java/com/solvd/selenium/components/FooterComponent.java
@@ -1,0 +1,18 @@
+package com.solvd.selenium.components;
+
+import org.openqa.selenium.SearchContext;
+import org.openqa.selenium.WebDriver;
+
+import com.zebrunner.carina.webdriver.gui.AbstractUIObject;
+
+public class FooterComponent extends AbstractUIObject {
+
+    // FooterComponent does not have any specific elements or methods defined yet.
+    // This class can be extended in the future to include footer-specific
+    // functionality.
+
+    public FooterComponent(WebDriver driver, SearchContext searchContext) {
+        super(driver, searchContext);
+    }
+
+}
