+++ src/main/java/com/solvd/selenium/components/StorefrontContentComponent.java
@@ -1,0 +1,19 @@
+package com.solvd.selenium.components;
+
+import org.openqa.selenium.SearchContext;
+import org.openqa.selenium.WebDriver;
+
+import com.zebrunner.carina.webdriver.gui.AbstractUIObject;
+
+public class StorefrontContentComponent extends AbstractUIObject {
+
+    // This class is currently empty and serves as a placeholder for future
+    // storefront content components.
+    // It can be extended to include specific functionality related to the
+    // storefront content.
+
+    public StorefrontContentComponent(WebDriver driver, SearchContext searchContext) {
+        super(driver, searchContext);
+    }
+
+}
