+++ src/main/java/com/solvd/selenium/pages/android/ProductPage.java
@@ -3,6 +3,7 @@ package com.solvd.selenium.pages.android;
 import org.openqa.selenium.WebDriver;
 
 import com.solvd.selenium.pages.common.ProductPageBase;
+import com.solvd.selenium.components.ProductComponent;
 import com.solvd.selenium.pages.common.BagPageBase;
 import com.zebrunner.carina.utils.factory.DeviceType;
 
@@ -14,79 +15,21 @@ public class ProductPage extends ProductPageBase {
     }
 
     @Override
+    public ProductComponent getProduct() {
+        // TODO: implement logic
+        return null;
+    }
+
+    @Override
     public BagPageBase clickViewBagButton() {
         // TODO: implement logic
         return null;
     }
 
     @Override
-    public void selectFirstAvailableSize() {
-        // TODO: implement logic
-    }
-
-    @Override
     public boolean isAddToBagConfirmationPresent() {
         // TODO: implement logic
         return false;
-    }
-
-    @Override
-    public boolean isProductNamePresent() {
-        // TODO: implement logic
-        return false;
-    }
-
-    @Override
-    public boolean isProductPricePresent() {
-        // TODO: implement logic
-        return false;
-    }
-
-    @Override
-    public boolean isProductDescriptionPresent() {
-        // TODO: implement logic
-        return false;
-    }
-
-    @Override
-    public boolean areAllProductImagesPresent() {
-        // TODO: implement logic
-        return false;
-    }
-
-    @Override
-    public boolean isSizeSelectorPresent() {
-        // TODO: implement logic
-        return false;
-    }
-
-    @Override
-    public boolean isColorSelectorPresent() {
-        // TODO: implement logic
-        return false;
-    }
-
-    @Override
-    public boolean isAddToBagButtonPresent() {
-        // TODO: implement logic
-        return false;
-    }
-
-    @Override
-    public boolean isFavouriteButtonPresent() {
-        // TODO: implement logic
-        return false;
-    }
-
-    @Override
-    public boolean isReviewsSectionPresent() {
-        // TODO: implement logic
-        return false;
-    }
-
-    @Override
-    public void clickAddToBag() {
-        // TODO: implement logic
     }
 
     @Override
