+++ src/main/java/com/solvd/selenium/pages/common/HomePageBase.java
@@ -1,11 +1,12 @@
 package com.solvd.selenium.pages.common;
 
-import java.util.List;
-
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.FindBy;
 
 import com.zebrunner.carina.webdriver.decorator.PageOpeningStrategy;
+import com.solvd.selenium.components.FooterComponent;
+import com.solvd.selenium.components.HeaderComponent;
+import com.solvd.selenium.components.StorefrontContentComponent;
 import com.zebrunner.carina.webdriver.decorator.ExtendedWebElement;
 import com.zebrunner.carina.webdriver.gui.AbstractPage;
 
@@ -25,33 +26,10 @@ public abstract class HomePageBase extends AbstractPage {
         acceptCookiesButton.clickIfPresent();
     }
 
-    public abstract SearchResultsPageBase clickSearchButton();
+    public abstract HeaderComponent getHeader();
 
-    public abstract SearchResultsPageBase searchForProduct(String searchText);
+    public abstract StorefrontContentComponent getStorefrontContent();
 
-    public abstract CategoryPageBase hoverOverMainCategoryAndClick(String category, String subCategory);
+    public abstract FooterComponent getFooter();
 
-    public abstract boolean isDeliveryTitlePresent();
-
-    public abstract String getDeliveryTitleText();
-
-    public abstract boolean isStoreLocatorPresent();
-
-    public abstract boolean isHelpLinkPresent();
-
-    public abstract boolean isLogoPresent();
-
-    public abstract boolean isSearchBarPresent();
-
-    public abstract boolean isAccountIconPresent();
-
-    public abstract boolean isFavoritesIconPresent();
-
-    public abstract boolean isShoppingBagIconPresent();
-
-    public abstract boolean isCheckoutButtonPresent();
-
-    public abstract boolean areAllMainMenuItemsPresent();
-
-    public abstract List<String> getMainMenuCategoriesList();
 }
