+++ src/main/resources/_config.properties
@@ -1,52 +1,48 @@
-#=================== Global configuration ============#
-url=https://www.next.co.uk
+######################################## Global configuration
+url=https://www.next.co.uk/
+
+######################################## Environment configuration
+env=DEMO
+DEMO.base=https://www.next.co.uk/
+
+######################################## DESKTOP MAC
 capabilities.platformName=MAC
+selenium_url=http://localhost:4444/
 capabilities.browserName=firefox
 capabilities.browserVersion=NULL
-selenium_url=http://localhost:4444
-custom_capabilities=NULL
-extra_capabilities=NULL
-app_version=NULL
 
-#desired thread count for parallel execution. If commented "thread-count" value from TestNG suite xml is used.
-#thread_count=1
+######################################## Android Local Mobile
+# capabilities.platformName=ANDROID
+# selenium_url=http://localhost:4723/
+# capabilities.automationName=uiautomator2
+# capabilities.deviceName=Medium_Phone_API_36.0
+# capabilities.udid=emulator-5554
+# capabilities.browserName=chrome
+# capabilities.noSign=true
+# capabilities.autoGrantPermissions=true
 
-#desired thread count for data provider(s). If commented "data-provider-thread-count" value from TestNG suite xml is used.
-#data_provider_thread_count=1
-
-#=====================================================#
-
-#=============== Environment configuration ===========#
-env=PROD
-#=====================================================#
-
-#============== Locale Configurations ================#
+######################################## Locale Configurations
 locale=en_US
 localization_testing=false
 localization_encoding=UTF-8
 language=en_US
-#=====================================================#
 
-#================ Retry configuration ================#
+######################################## Retry configuration
 #number of extra retries to execute tests until pass
 retry_count=0
-#=====================================================#
 
-#============== WebDriver configuration ==============#
+######################################## WebDriver configuration
 explicit_timeout=20
 disappear_timeout=2
 #number of extra retries to start driver session and interval in seconds
 init_retry_count=0
 init_retry_interval=1
-#=====================================================#
 
-#================ Report configuration ===============#
+######################################## Report configuration
 core_log_level=INFO
 auto_screenshot=true
 project_report_directory=./reports
 report_url=NULL
 max_screen_history=10
-
 email_list=NULL
 suite_name=NULL
-#=====================================================#
